fix(sonner): rename components.sooner package to components.sonner
Every type is named Sonner* but lived in a misspelled 'sooner' package.
Renamed the directory and updated package declarations and all imports in
the library and demo app before the public API locks it in. BREAKING for
any external consumer importing com.komoui.components.sooner.*.

Closes shadcn-ui-kmp-sun
