+++ komoui/src/commonMain/kotlin/com/komoui/components/sonner/SonnerProvider.kt
@@ -1,4 +1,4 @@
-package com.komoui.components.sooner
+package com.komoui.components.sonner
 
 import androidx.compose.material3.SnackbarDuration
 import kotlinx.coroutines.channels.Channel
